add documentation and option --dry-run #247

+++ src/cli/class-tainacan-cli-collection.php
@@ -10,6 +10,7 @@ class Cli_Collection {
 	private $collection_repository;
 	private $items_repository;
 	private $result_count;
+	private $dry_run = false;
 
 	public function __construct() {
 		$this->items_repository = Repositories\Items::get_instance();
@@ -17,6 +18,22 @@ class Cli_Collection {
 		$this->result_count = ['items' => 0, 'attachments' => 0];
 	}
 
+	/**
+	 * Show a list of collections.
+	 *
+	 * ## EXAMPLES
+   *
+	 * wp tainacan collection list
+	 * +------+-------------------+
+	 * | ID   | title             |
+	 * +------+-------------------+
+	 * | 1919 | Collection test 1 |
+	 * | 1201 | Collection test 1 |
+	 * | 1177 | Livros            |
+	 * | 1157 | autores           |
+	 * +------+-------------------+
+   *
+	 */
 	public function list() {
 		$response = [];
 		$collections = $this->collection_repository->fetch(['per-page'=>-1], 'OBJECT');
@@ -26,6 +43,30 @@ class Cli_Collection {
 		\WP_CLI\Utils\format_items( 'table', $response, ['ID', 'title'] );
 	}
 
+	/**
+	 * remove items of specific collection.
+	 *
+	 * ## OPTIONS
+	 * [--collection-id=<value>]
+	 * : <value> ID of specifies the collection that will have your items removed.
+	 * 
+	 * [--permanently]
+   * : if items do not need to be trashed, if present permanently deletes items.
+	 * 
+	 * [--dry-run]
+	 * : only count the total of item which will remove, just output a report 
+	 * 
+	 * ## EXAMPLES
+	 * 
+	 * wp tainacan collection clean --collection-id=1201 --permanently
+	 * 
+	 * cleaning collection items 
+	 * 100% [============================================================================================] 0:00 / 0:00
+	 * Success: 
+	 * 10 items removed
+	 * 23 attachments removed
+	 * 
+	 */
 	public function clean($args, $assoc_args) {
 		$permanently = false;
 		if( empty($assoc_args['collection-id']) ) {
@@ -37,17 +78,23 @@ class Cli_Collection {
 			$permanently = true; 
 		}
 
+		$this->dry_run = false;
+		if ( !empty($assoc_args['dry-run']) ) {
+			$this->dry_run = true;
+		}
+
 		$per_page = 50; $page = 0;
 		$args = [
 			'posts_per_page'=> $per_page,
 			'paged' => $page,
-			'post_status' => array('publish', 'pending', 'draft', 'auto-draft', 'future', 'private', 'inherit', 'trash')
+			'post_status' => get_post_stati()
+			//'post_status' => array('publish', 'pending', 'draft', 'auto-draft', 'future', 'private', 'inherit', 'trash')
 		];
 		$collection_items = $this->items_repository->fetch($args, $collection_id, 'WP_Query');
 		$total = $collection_items->found_posts;
 		$last_page = ceil($total/$per_page);
 
-		$progress = \WP_CLI\Utils\make_progress_bar( "\ncleaning collection items: \n", $total );
+		$progress = \WP_CLI\Utils\make_progress_bar( "cleaning collection items:", $total );
 		while ($page <= $last_page) {
 			if ($collection_items->have_posts()) {
 				while ( $collection_items->have_posts() ) {
@@ -62,13 +109,17 @@ class Cli_Collection {
 		}
 		$progress->finish();
 
-		$msg = "\n" . $this->result_count['items']       . " items removed" .
-					 "\n" .	$this->result_count['attachments'] . " attachments removed";
+		if ($permanently) {
+			$msg = "\n" . $this->result_count['items']       . " items removed" .
+						 "\n" .	$this->result_count['attachments'] . " attachments removed";
+		} else {
+			$msg = "\n" . $this->result_count['items']       . " items moved to trash";
+		}
 
 		\WP_CLI::success( $msg );
 	}
 
-	private function delete_item( $item, $permanently = false ) {
+	private function delete_item( $item, $permanently = false) {
 		if (! $item instanceof Entities\Item) {
 			\WP_CLI::error( 'An item with this ID was not found', true );
 		}
@@ -76,15 +127,18 @@ class Cli_Collection {
 		$this->result_count['items']++;
 		if($permanently == true) {
 			$this->delete_attachments($item);
-			$item = $this->items_repository->delete($item);
+			if(!$this->dry_run) {
+				$item = $this->items_repository->delete($item);
+			}
 		} else {
-			$item = $this->items_repository->trash($item);
+			if(!$this->dry_run) {
+				$item = $this->items_repository->trash($item);
+			}
 		}
 		return $item;
 	}
 
 	private function delete_attachments ( $item ) {
-		$this->result_count['attachments']++;
 		$attachment_list = array_values(
 			get_children(
 				array(
@@ -96,7 +150,10 @@ class Cli_Collection {
 			)
 		);
 		foreach ($attachment_list as $attachment) {
-			wp_delete_attachment($attachment->ID);
+			$this->result_count['attachments']++;
+			if(!$this->dry_run) {
+				wp_delete_attachment($attachment->ID);
+			}
 		}
 	}
